Added mobile_session_id to elements/sessions retrieval call for CustomerSheet (#5479)
## Summary
<!-- Simple summary of what was changed. -->
Added mobile_session_id as a param on the elements/sessions call for
CustomerSheet. Was previously added for PaymentSheet but forgot about
CustomerSheet
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
https://jira.corp.stripe.com/browse/MOBILESDK-4025
## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
Network stubs
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A