Tweak Link UI for iOS 26 (#5448)
## Summary
<!-- Simple summary of what was changed. -->

This pull request updates the native Link experience and the inline
signup for iOS 26. We aren't doing a big redesign at this time, but are
focusing on making sure it doesn't look out of place.

Just like the MPE experience, this change is gated behind
`LiquidGlassDetector.isEnabled`.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

Liquid 🥛.

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

Snapshots in https://github.com/stripe/stripe-ios/pull/5474.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->