Opt-in API for iOS26 changes behind dev flag & impl for navbar opt-in (#5507)
## Summary
Introduces an opt-in flag for iOS26 changes behind a dev flag

## Motivation
We need a way to start removing `LiquidGlassDesign.isEnabled` from all
parts of our code base. The idea would be to merge this in, then we can
start removing parts of our codebase that rely on
`LiquidGlassDesign.isEnabled`.

## Testing
Updated existing snapshot test code.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->