Fix divider visibility in `MultiElementRow` (#5530)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes an issue where dividers between elements could
stay visible in `MultiElementRow` even if the adjacent elements are set
to be hidden.

~We now only store the actual views (and not the dividers) and
re-compute the entire list of `arrangedSubviews` whenever an element is
toggled.~

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

Polish.

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

Updated snapshots.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->