[iOS 26] Rework readonly component background color (#5580)
## Summary
Reworked the readonly component to not depend on LiquidGlass.isEnabled
and instead depend on the brightness of the background color.

This also fixes an existing bug where the color is brightened/darkened
according to the device dark mode setting rather than the brightness of
the background - incorrect if e.g. the user has set a black background
for light and dark mode.

## Motivation
iOS 26

## Testing
Existing snapshot tests


https://github.com/user-attachments/assets/c52dab35-630a-41dd-8314-b9d2c315edc3



https://github.com/user-attachments/assets/0a4d0a81-7404-49c7-9020-c9d28e2b4c95