Fix issue with invalid Link selection for RUX flows (#5592)
## Summary
<!-- Simple summary of what was changed. -->

This pull request resolves an issue where dropping out of the Link
bottom sheet and back into MPE would leave `link(confirmOption:
.wallet)` as the currently selected payment option, which isn't a great
experience. Instead, we now clear the current selection.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->