Adjust when hcaptcha token fetched (#5527)
## Summary
<!-- Simple summary of what was changed. -->
Instead of fetching the token before calling confirm from PaymentSheet,
FlowController. Embedded, CustomerSheet regardless of payment method and
passing that around, we pass the PassiveCaptchaChallenge and call
fetchToken only when it's needed.
Omit it for ApplePay and saved pms.
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
First captcha load takes several seconds. This disrupts the user
experience for quick payment methods that don't require users to input
payment details (like saved pms, ApplePay). As captcha is meant to guard
against card testing, neither of these cases need passive captcha.
Discussion on Slack
[here](https://stripe.slack.com/archives/C08SZ67N0NM/p1757615895108549)

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
Manual testing, CI. Tried to add network tests in
StripePaymentSheetTests, but HCaptcha validation doesn't seem to work in
test environments without a window (Webview can't load)
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A