[WalletButtonsView] Don't unset Apple Pay/Link on FlowController update (#5639)
## Summary
I don't love the state management issues here, but this should fix the
immediate problem. Just check if we have a previousPaymentOption (only
true in an update call) and skip the "unset the default PM if it's Apple
Pay or Link" behavior.

## Motivation
Link/Apple Pay get dropped after an `update` call.

## Testing
PaymentSheet Example

## Changelog
N/A