remove layout margins pinning
In this stack view, we were pining the payment method view to the leading and trailling anchors of the stack view, as well as setting this layout margin pinning. These two conflict and when the layout system resolved constraints, it resolved to removing these layout margin pinning constrinats. So let’s explicitly remove them, since that’s how our view was rendering anyway.