Update Link recollection UI (#5575)
## Summary
<!-- Simple summary of what was changed. -->

This pull request updates the UI for recollecting expiry date and CVC in
the Link wallet.

We no longer show an error message, but simply present the text fields
for recollection. We do keep the error icon for expired cards in the
payment method picker.

We also tweak the spacing between sections.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

Design feedback.

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

Updated snapshots.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->