Error Message Styling for Link Payment View (#5643)
## Summary
Updates the styling of error messages presented by
`PayWithLinkViewController`. This is accomplished by adding a style to
`LinkHintMessageView` for `error`s along with new properties for icon
and color.

## Motivation
The previous UI didn't match Android and it didn't match the design
language used by the rest of Link.

[Here's](https://github.com/user-attachments/assets/20d8bc6d-ff41-45df-b06f-c02966e1479b)
what the old UI looked like.

## Testing
Snapshot tests were added along with a `#Preview` definition to help
with future tweaks.

---------

Co-authored-by: Mat Schmid <mats@stripe.com>