Allow overriding country code in test mode (#5626)
## Summary
<!-- Simple summary of what was changed. -->

This pull request updates the usage of `userOverrideCountry` in the
playground. Instead of only using it locally, we now send it to
`elements/sessions`, which allows us consider the override country even
in fields that are being computed in the backend.

See
[here](https://livegrep.corp.stripe.com/view/stripe-internal/pay-server/lib/elements/command/retrieve_session_info.rb#L1306)
for the implementation of `country_override` on the backend.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

Easier testing for non-US folks.

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->