Fix Xcode 15 Link New Payment Snapshot Tests (#5669)
## Summary
The tests on Xcode 15
[failed](https://app.bitrise.io/app/b220f4ba85d134a7/pipelines/52a67dd3-86ca-4485-886a-cdcbe0c28ea7?tests_filter_status=all)
when deploying v24.24.2. This PR fixes the build error with these tests
by removing `-` and `_` characters from the file name of the test. These
tests were not getting skipped based on the [file name
regex](https://stripe.sourcegraphcloud.com/stripe/stripe-ios/-/blob/ci_scripts/test.rb?L90-94)
when they were supposed to.

This PR also adds a missing `// @iOS26` tag to the tests and includes
the test in the updated test plan.

## Motivation
Fix failing tests.

## Testing
Included the legacy tests in a bitrise build and
[passed](https://app.bitrise.io/app/b220f4ba85d134a7/pipelines/50c82fdd-032c-4f1b-8839-a28c3672e4a9?tests_filter_status=all)!