Update logic for isEnabledInMerchantApp (#5683)
## Summary
We detected that when compiling with xcode26, but executing on <iOS26
target, the done button becomes unresponsive. But this applies to a
broader set of problems where we may be executing code that was intended
for liquid glass to execute on devices < iOS26.

## Motivation
Fix regression

## Testing
Manually

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->

---------

Co-authored-by: Yuki <yuki@stripe.com>