Prevent hcaptcha crash (#5625)
## Summary
<!-- Simple summary of what was changed. -->
Refactored PassiveCaptchaChallenge to use error throwing pattern instead
of Result and refactored the functions. Instead of `start`, we have
`fetchToken` that fetches the token or throws, and instead of the
original `fetchToken`, we have `fetchTokenWithTimeout`, which awaits
`fetchToken` with a timeout.
Added a lock with the nillableContinuation in `fetchToken` to prevent
any possible Swift Task continuation misuse crash if it somehow tried to
resume continuation from another thread.
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
assertionFailure: reported crashes due to a race condition that led to
an assertionFailure due to validationTask not being initialized by the
time await validationTask?.value was called. The refactor should fix
this.
Swift Task continuation misuse crash: reported crashes due to attempted
continuation resume.
https://jira.corp.stripe.com/browse/MOBILESDK-4057
## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
Manual, CI
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A