Add `disableFundingSources` to Link configuration (#5556)
## Summary
<!-- Simple summary of what was changed. -->

This pull request adds a new field `disallowFundingSourceCreation` to
`PaymentSheet.LinkConfiguration`. It allows merchants to dynamically
toggle the visibility of Link funding sources, e.g. the bank tab.

cc @ericazhou-stripe

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

[APIREVIEW-4236](https://jira.corp.stripe.com/browse/APIREVIEW-4236)

## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->

Unit tests.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->