Add client_attribution_metadata to payment details and adjust for LinkController (#5684)
## Summary
<!-- Simple summary of what was changed. -->
Added `client_attribution_metadata` to the
`/v1/consumers/payment_details` endpoint. Added a
PaymentSheetAnalyticsHelper integration shape, .linkController, because
LinkController is not a part of MPE and therefore should not send
`client_attribution_metadata`. Added `client_attribution_metadata` to
financial connections Link calls
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
`/v1/consumers/payment_details` has come into scope for Project Weave
## Testing
<!-- How was the code tested? Be as specific as possible. -->
<!-- Ignored Tests: Did you newly ignore a test in this PR? If so,
please open an R4 incident so that the test can be re-enabled as soon as
possible-->
CI
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A