Make analytics report timestamp (#5716)
## Motivation
We need to send a client-side timestamp. Can’t rely on created
timestamps since that’s when it hits the server + batching effects.
There’s too much variance to be useful for tracking time between two
events.

See created note in analytics.elements_events:

> “The server-side timestamp the event was received by AEL. Note that
due to event batching, this is not a dependable way to determine what
order events were emitted in, or when exactly they were emitted. See
event_count for ordering, and timestamp for the time the event was
created in the browser.”

and its ‘timestamp’ column

> “The client-side timestamp the event was created at in the browser.”

## Testing
See test

## Changelog
Not user facing